Add OpenLayers marker filtering and coloring
